requirements.txt
django==4.2.11
mysqlclient==2.2.4
transformers==4.45.1
torch==2.4.1
faiss-cpu==1.9.0
numpy==1.26.4
python-dotenv==1.0.1